complete the crud operations for user entity

+++ src/main/java/com/spring/Crudproject/Service/UserService.java
@@ -1,14 +1,19 @@
 package com.spring.Crudproject.Service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.spring.Crudproject.Exception.RoleHandler;
 import com.spring.Crudproject.Model.Response;
 import com.spring.Crudproject.Model.Role;
 import com.spring.Crudproject.Model.User;
 import com.spring.Crudproject.Repository.RoleRepository;
 import com.spring.Crudproject.Repository.UserRepository;
 import lombok.AllArgsConstructor;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ReflectionUtils;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.lang.reflect.Field;
@@ -17,6 +22,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -26,10 +32,9 @@ public class UserService {
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
 
-    public Response ShowUsers(){
+    public ResponseEntity<Object> ShowUsers(){
         List<User> x=userRepository.findAll();
-
-        return x.isEmpty()? new Response<String>("no user exist !!"):new Response<List<User>>(x);
+        return x.isEmpty()?new ResponseEntity<Object>("no user exist !!",HttpStatus.BAD_REQUEST):new ResponseEntity<Object>(x,HttpStatus.OK);
     }
 
     public User ShowUser(Integer id){
@@ -42,45 +47,44 @@ public class UserService {
         });
     }
 
-    public void SaveUser(Map<String, Object> requestBody) throws Exception{
+    public User SaveUser(User user) throws RoleHandler{
 
-        List<String> fl=new ArrayList<String>();
-        Field[] f = User.class.getDeclaredFields();
-        for(Field x : f){
-            if(x.getName()!="id")
-                fl.add(x.getName());
-        }
-
-
-
-        for(String x : fl){
-            if(!requestBody.keySet().contains(x)){
-                throw new Exception(x+" not found !!");
-            }
-        }
-
-        for(String x : requestBody.keySet()){
-            if(!fl.contains(x)){
-                throw new Exception(x+" not wanted field !!");
-            }
-        }
-
-        User user=new ObjectMapper().convertValue(requestBody, User.class);
-
-        Optional<User> x=userRepository.findByEmailOrUserName(user.getEmail(),user.getUserName());
+        // List<String> fl=new ArrayList<String>();
+        // Field[] f = User.class.getDeclaredFields();
+        // for(Field x : f){
+        //     if(x.getName()!="id")
+        //         fl.add(x.getName());
+        // }
 
 
-        if(x.isPresent()){
-            throw new Exception("user already exist !");
-        }
+
+        // for(String x : fl){
+        //     if(!requestBody.keySet().contains(x)){
+        //         throw new Exception(x+" not found !!");
+        //     }
+        // }
+
+        // for(String x : requestBody.keySet()){
+        //     if(!fl.contains(x)){
+        //         throw new Exception(x+" not wanted field !!");
+        //     }
+        // }
+
+        // User user=new ObjectMapper().convertValue(requestBody, User.class);
+
+        // Optional<User> x=userRepository.findByEmailOrUserName(user.getEmail(),user.getUserName());
+
+
+        // if(x.isPresent()){
+        //     throw new Exception("user already exist !");
+        // }
 
         Optional<Role> r= roleRepository.findById(user.getRole().getRole_id());
         if(!r.isPresent()){
-            throw new Exception("role id not exist !!");
+            throw new RoleHandler("role id not exist !!");
         }
-
         
-        userRepository.save(user);
+        return userRepository.save(user);
 
 
         
@@ -102,6 +106,9 @@ public class UserService {
     }
 
     public void DeleteUser(Integer id){
+        if(!userRepository.findById(id).isPresent())
+            throw new NoSuchElementException();
+            
         userRepository.deleteById(id);
     }
 }
